refactor contactRoures
