+++ models/contacts/index.js
@@ -2,6 +2,4 @@
 
 const Contact = require('./contactModel');
 
-module.exports = {
-  Contact,
-};
+module.exports = Contact;
