+++ routes/contacts/index.js
@@ -1,0 +1,5 @@
+// ./routes/contacts/index.js
+
+const routes = require('./contactRoutes');
+
+module.exports = routes;
