+++ services/contacts/checkContactExists.js
@@ -1,6 +1,6 @@
 // ./services/contacts/createContact.js
 
-const { Contact } = require('../../models');
+const Contact = require('../../models/contacts');
 
 /**
  * Перевірка, чи існує контакт.
