+++ services/contacts/createContact.js
@@ -1,6 +1,6 @@
 // ./services/contacts/createContact.js
 
-const { Contact } = require('../../models');
+const Contact = require('../../models/contacts');
 
 /**
  * Створити новий контакт.
