+++ services/contacts/getOneContact.js
@@ -1,6 +1,6 @@
 // ./services/contacts/getOneContact.js
 
-const { Contact } = require('../../models');
+const Contact = require('../../models/contacts');
 
 /**
  * Знайти контакт за ідентифікатором.
